requirements.txt
mysql-connector-python==8.0.20
numpy==1.19.0
pandas==1.0.5
protobuf==3.12.2
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
